2437: add useDatagridSelection hook and use it on annotation and parsingResults datagrid
